loop_overlay2 driver enabling stateless handling of images
Uses loopback devices for layers that are stacked for the
container using overlayfs. This enables support for CoW-like
image layers on a shared file system that lacks overlayfs support.

Driver also backs the image and layer stores using loopback devices
that can reside on shared storage. On graphdriver initialization,
stores found in the root or fallback directories are merged together
using overlayfs, or if only one of the two directories are writable,
then the loopback device is bind mounted. This means changes to the
stores are written and persisted to a loopback device, which is
chosen in order of preference:
  1) loopback file in root directory
  2) loopback file in fallback directory

Additionaly using a loopback device on shared and read-only storage
as the lower dir in an overlay mount provides a consistent view of
already available images across multiple systems. These systems have
"stateless" image stores in the sense that the Docker daemon could
operate with no persistent local storage. The RW working layer of
containers will not be persisted unless an explicit 'docker commit'
is made by the user.

Much of the code is based on the overlay2 driver.